Move count_flops into gem/utils.py
